+++ aurora/src/main/java/com/trishla/aurora/post/foundDog/dto/FoundDog.java
@@ -14,7 +14,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class FoundDog {
-    private int UID;
+    private long UID;
     private String breed;
     private Gender gender;
     private DogPhysicalAttributes distinctiveFeatures;
